refactoring: socket.io RoomSocket
