+++ backend/node/socket_io_server/RoomSocket.js
@@ -2,14 +2,47 @@ import {
 	SOCKET_IO_EVENT_JOIN_ROOM,
 	SOCKET_IO_EVENT_LEAVE_ROOM,
 } from "../constants/socket.io-Events.js";
+import logger from "./logger.js";
 
 // todo test
 class RoomSocket {
 	constructor({socket, server, handlerEventPair}) {
+		this.id = socket.id;
 		this.socket = socket;
 		this.server = server;
 		this.handlerPair = handlerEventPair;
 		this.registeredHandler = [];
+		this.room = null;
+
+		this.addRoomHandler();
+	}
+
+	addRoomHandler() {
+		const onJoinRoom = async req => {
+			const {room} = req;
+
+			if (this.room) {
+				logger.error(`id: ${this.id} already join in room`);
+				return;
+			}
+
+			try {
+				this.joinRoom(room);
+			} catch (e) {
+				logger.error(`error raised while join room ${e}`);
+			}
+		};
+
+		const onLeaveRoom = async () => {
+			try {
+				this.leaveRoom();
+			} catch (e) {
+				logger.error(`error raised while leave room ${e}`);
+			}
+		};
+
+		this.socket.on(SOCKET_IO_EVENT_JOIN_ROOM, onJoinRoom);
+		this.socket.on(SOCKET_IO_EVENT_LEAVE_ROOM, onLeaveRoom);
 	}
 
 	joinRoom(room) {
@@ -43,7 +76,11 @@ class RoomSocket {
 		};
 
 		const wrappedSocketHandler = data => {
-			handler(data, roomEmit, {socket: this.socket, server: this.server});
+			handler(data, roomEmit, {
+				socket: this.socket,
+				server: this.server,
+				room: this.room,
+			});
 		};
 
 		this.socket.on(event, wrappedSocketHandler);
